Starting update to program list view.
